fix: Compile Issue
- CS0266 Cannot implicitly convert type 'int' to 'TextMateSharp.Themes.FontStyle'
- CS0019 Operator '!=' cannot be applied to operands of type 'FontStyle' and 'int'
- CS0019 Operator '&' cannot be applied to operands of type 'FontStyle' and 'int'

+++ src/AvaloniaEdit.TextMate/TextMateColoringTransformer.cs
@@ -175,7 +175,7 @@ namespace AvaloniaEdit.TextMate
                         background = themeRule.background;
 
                     if (fontStyle == 0 && themeRule.fontStyle > 0)
-                        fontStyle = (int)themeRule.fontStyle;
+                        fontStyle = themeRule.fontStyle;
                 }
 
                 if (transformations[i] == null)
